[AIP-21] Move contrib.hooks.azure_container_registry_hook airflow.contrib.hooks.azure_container_registry_hook
