+++ tests/providers/microsoft/azure/hooks/test_azure_container_registry.py
@@ -19,8 +19,8 @@
 
 import unittest
 
-from airflow.contrib.hooks.azure_container_registry_hook import AzureContainerRegistryHook
 from airflow.models import Connection
+from airflow.providers.microsoft.azure.hooks.azure_container_registry import AzureContainerRegistryHook
 from airflow.utils import db
